uplift/bot: Refactor Bugzilla API calls into a separate file.
